pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-video-agent"
version = "2.0.0"
description = "Topic/Script → Voiceover → Images → Video → YouTube — fully automated AI video pipeline."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Anand Singh"},
]
keywords = ["AI", "video", "automation", "youtube", "elevenlabs", "openai", "dall-e"]

dependencies = [
    "elevenlabs>=1.0.0",
    "openai>=1.12.0",
    "moviepy>=1.0.3",
    "Pillow>=10.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "httpx>=0.25.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.3.0",
]

[project.scripts]
ai-video-agent = "ai_video_agent.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
